trust_indicator/manipulationdet/requirements.txt
albumentations==1.3.1
certifi==2022.12.7
charset-normalizer==3.3.2
colorama==0.4.6
Faker==18.13.0
filelock==3.12.2
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
joblib==1.3.2
networkx==2.6.3
numpy==1.21.6
opencv-contrib-python==3.4.3.18
opencv-python==3.4.3.18
opencv-python-headless==4.9.0.80
packaging==24.0
Pillow==9.5.0
pip==22.3.1
protobuf==4.24.4
python-dateutil==2.9.0.post0
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.12
torch==1.8.1
torchvision==0.9.1
tqdm==4.66.2
typing_extensions==4.7.1
urllib3==2.0.7
wheel==0.38.4
wincertstore==0.2
zipp==3.15.0
